backend_2.0/pyproject.toml
[project]
name = "backend-2-0"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask[async]>=3.1.1",
    "flask-cors>=6.0.1",
    "google-adk>=1.6.1",
    "python-dotenv>=1.1.1",
]
